rustc_metadata: Move some code around
Plugin search doesn't need a crate loader, only crate locator
